+++ lib/rubocop/runner.rb
@@ -24,6 +24,14 @@ module RuboCop
 
     def initialize(options, config_store)
       @options = options
+
+      if @options.key?(:rails)
+        warn <<~MESSAGE
+          `-R/--rails` option and Rails cops will be removed from RuboCop 0.72. Use the `rubocop-rails` gem instead.
+          https://github.com/rubocop-hq/rubocop/blob/master/manual/migrate_rails_cops.md
+        MESSAGE
+      end
+
       @config_store = config_store
       @errors = []
       @warnings = []
@@ -273,7 +281,10 @@ module RuboCop
 
     def inspect_file(processed_source)
       config = @config_store.for(processed_source.path)
-      enable_rails_cops(config) if @options[:rails]
+      if @options[:rails] ||
+         ConfigLoader.required_features.include?('rubocop-rails')
+        enable_rails_cops(config)
+      end
       team = Cop::Team.new(mobilized_cop_classes(config), config, @options)
       offenses = team.inspect_file(processed_source)
       @errors.concat(team.errors)